# src/backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Specvital API
  description: Test specification analysis service for GitHub repositories
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/spec-view/{analysisId}:
    parameters:
      - name: analysisId
        in: path
        required: true
        description: Analysis ID (UUID) to fetch spec document for
        schema:
          type: string
          format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
    get:
      operationId: getSpecDocument
      summary: Get specification document for analysis
      description: |
        Retrieves the AI-generated specification document for a given analysis.
        Returns full document hierarchy: domains → features → behaviors.
        If document is being generated, returns status with progress.
        Optionally filter by language and/or version parameter.
      tags:
        - Spec View
      security:
        - cookieAuth: []
        - {}
      parameters:
        - name: language
          in: query
          required: false
          description: Filter by language. If not specified, returns the most recent document.
          schema:
            $ref: "#/components/schemas/SpecLanguage"
        - name: version
          in: query
          required: false
          description: Specific version number to retrieve. Requires language parameter. If not specified, returns the latest version.
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        "200":
          description: Spec document retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecDocumentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/spec-view/{analysisId}/versions:
    parameters:
      - name: analysisId
        in: path
        required: true
        description: Analysis ID (UUID) to fetch version history for
        schema:
          type: string
          format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
    get:
      operationId: getSpecVersions
      summary: Get version history for a specific language
      description: |
        Returns all versions of the spec document for a given analysis and language.
        Ordered by version number descending (latest first).
      tags:
        - Spec View
      security:
        - cookieAuth: []
        - {}
      parameters:
        - name: language
          in: query
          required: true
          description: Language to get version history for
          schema:
            $ref: "#/components/schemas/SpecLanguage"
      responses:
        "200":
          description: Version history retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionHistoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/spec-view/generate:
    post:
      operationId: requestSpecGeneration
      summary: Request spec document generation
      description: |
        Requests generation of a specification document for an analysis.
        Enqueues a River job for async AI processing.
        Returns immediately with job status.
      tags:
        - Spec View
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestSpecGenerationRequest"
      responses:
        "202":
          description: Generation request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestSpecGenerationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Generation already in progress or completed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/spec-view/status/{analysisId}:
    parameters:
      - name: analysisId
        in: path
        required: true
        description: Analysis ID (UUID) to check generation status for
        schema:
          type: string
          format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
    get:
      operationId: getSpecGenerationStatus
      summary: Get spec generation status
      description: |
        Returns the current status of spec document generation.
        Useful for polling during generation.
        If language is specified, returns status for that specific language.
      tags:
        - Spec View
      security:
        - cookieAuth: []
        - {}
      parameters:
        - name: language
          in: query
          required: false
          description: Language to check status for (e.g., English, Korean)
          schema:
            type: string
          example: English
      responses:
        "200":
          description: Generation status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecGenerationStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # User Subscription APIs
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT token stored in HTTP-only cookie
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Forbidden:
      description: Access forbidden
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
...
    TestStatus:
      type: string
      enum:
        - active
        - focused
        - skipped
        - todo
        - xfail
      description: |
        Test status indicator:
        - active: Normal test that will run
        - focused: Test marked to run exclusively (e.g., it.only)
        - skipped: Test marked to be skipped (e.g., it.skip)
        - todo: Placeholder test to be implemented
        - xfail: Expected to fail (pytest xfail)

    # Framework
    Framework:
      type: string
      description: Testing framework identifier
      example: vitest

    # Summary
...
    ProblemDetail:
      type: object
      required:
        - status
        - title
        - detail
      properties:
        type:
          type: string
          format: uri
          default: about:blank
          description: URI identifying the problem type
        title:
          type: string
          description: Human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          description: URI reference identifying the specific occurrence
        rateLimit:
          $ref: "#/components/schemas/RateLimitInfo"

    # Rate Limit Info
    RateLimitInfo:
      type: object
      required:
        - limit
        - remaining
        - resetAt
      properties:
        limit:
          type: integer
          description: Maximum requests allowed per time window
        remaining:
          type: integer
          description: Remaining requests in current window
        resetAt:
          type: integer
          format: int64
          description: Unix timestamp when the rate limit resets

    # Auth Schemas
...
    SpecDocumentResponse:
      oneOf:
        - $ref: "#/components/schemas/SpecDocumentCompleted"
        - $ref: "#/components/schemas/SpecDocumentGenerating"
      discriminator:
        propertyName: status
        mapping:
          completed: "#/components/schemas/SpecDocumentCompleted"
          generating: "#/components/schemas/SpecDocumentGenerating"
    SpecDocumentCompleted:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          const: completed
        data:
          $ref: "#/components/schemas/SpecDocument"
    SpecDocumentGenerating:
      type: object
      required:
        - status
        - generationStatus
      properties:
        status:
          type: string
          const: generating
        generationStatus:
          $ref: "#/components/schemas/SpecGenerationStatus"
    SpecDocument:
      type: object
      required:
        - id
        - analysisId
        - language
        - version
        - domains
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Spec document ID
        analysisId:
          type: string
          format: uuid
          description: Associated analysis ID
        language:
          $ref: "#/components/schemas/SpecLanguage"
        version:
          type: integer
          minimum: 1
          description: Document version number (increments on regeneration)
          example: 1
        executiveSummary:
          type: string
          description: AI-generated executive summary of the test suite
          example: "This repository contains a comprehensive authentication system with 45 test behaviors covering user login, session management, and password recovery."
        modelId:
          type: string
          description: AI model ID used for generation
          example: "gemini-2.0-flash"
        availableLanguages:
          type: array
          items:
            $ref: "#/components/schemas/AvailableLanguageInfo"
          description: List of all available languages for this analysis
        domains:
          type: array
          items:
            $ref: "#/components/schemas/SpecDomain"
          description: Domain classifications
        createdAt:
          type: string
          format: date-time
          description: Document creation timestamp
    AvailableLanguageInfo:
      type: object
      required:
        - language
        - latestVersion
        - createdAt
      properties:
        language:
          $ref: "#/components/schemas/SpecLanguage"
        latestVersion:
          type: integer
          minimum: 1
          description: Latest version number for this language
          example: 3
        createdAt:
          type: string
          format: date-time
          description: When this language was last generated
    VersionInfo:
      type: object
      required:
        - version
        - createdAt
      properties:
        version:
          type: integer
          minimum: 1
          description: Version number
          example: 1
        createdAt:
          type: string
          format: date-time
          description: When this version was created
        modelId:
          type: string
          description: AI model ID used for this version
          example: "gemini-2.0-flash"
    VersionHistoryResponse:
      type: object
      required:
        - data
        - language
        - latestVersion
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/VersionInfo"
          description: List of versions ordered by version number descending
        language:
          $ref: "#/components/schemas/SpecLanguage"
        latestVersion:
          type: integer
          minimum: 1
          description: The latest version number for this language
          example: 3
    SpecDomain:
      type: object
      required:
        - id
        - name
        - sortOrder
        - features
      properties:
        id:
          type: string
          format: uuid
          description: Domain ID
        name:
          type: string
          description: Domain name
          example: "Authentication"
        description:
          type: string
          description: Domain description
          example: "User authentication and session management"
        classificationConfidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: AI classification confidence score (0-1)
          example: 0.92
        sortOrder:
          type: integer
          minimum: 0
          description: Display order within document
        features:
          type: array
          items:
            $ref: "#/components/schemas/SpecFeature"
          description: Features within this domain
    SpecFeature:
      type: object
      required:
        - id
        - name
        - sortOrder
        - behaviors
      properties:
        id:
          type: string
          format: uuid
          description: Feature ID
        name:
          type: string
          description: Feature name
          example: "Login"
        description:
          type: string
          description: Feature description
          example: "User login with various authentication methods"
        sortOrder:
          type: integer
          minimum: 0
          description: Display order within domain
        behaviors:
          type: array
          items:
            $ref: "#/components/schemas/SpecBehavior"
          description: Behaviors (converted test cases) within this feature
    SpecBehavior:
      type: object
      required:
        - id
        - originalName
        - convertedDescription
        - sortOrder
      properties:
        id:
          type: string
          format: uuid
          description: Behavior ID
        originalName:
          type: string
          description: Original test case name
          example: "should create session when credentials are valid"
        convertedDescription:
          type: string
          description: AI-converted natural language description
          example: "Creates a new session when valid credentials are provided"
        sortOrder:
          type: integer
          minimum: 0
          description: Display order within feature
        sourceTestCaseId:
          type: string
          format: uuid
          description: Reference to original test case
        sourceInfo:
          $ref: "#/components/schemas/SpecBehaviorSourceInfo"
    SpecBehaviorSourceInfo:
      type: object
      required:
        - filePath
        - lineNumber
        - status
        - framework
      properties:
        filePath:
          type: string
          description: Test file path
          example: "src/__tests__/auth.test.ts"
        lineNumber:
          type: integer
          minimum: 1
          description: Line number in the test file
        status:
          $ref: "#/components/schemas/TestStatus"
        framework:
          $ref: "#/components/schemas/Framework"
    SpecLanguage:
      type: string
      enum:
        - Arabic
        - Chinese
        - Czech
        - Danish
        - Dutch
        - English
        - Finnish
        - French
        - German
        - Greek
        - Hindi
        - Indonesian
        - Italian
        - Japanese
        - Korean
        - Polish
        - Portuguese
        - Russian
        - Spanish
        - Swedish
        - Thai
        - Turkish
        - Ukrainian
        - Vietnamese
      default: English
      description: Target language for spec document generation (24 languages supported)
    RequestSpecGenerationRequest:
      type: object
      required:
        - analysisId
      properties:
        analysisId:
          type: string
          format: uuid
          description: Analysis ID to generate spec document for
        language:
          $ref: "#/components/schemas/SpecLanguage"
        isForceRegenerate:
          type: boolean
          default: false
          description: Force regeneration even if document exists
    RequestSpecGenerationResponse:
      type: object
      required:
        - status
        - analysisId
      properties:
        status:
          $ref: "#/components/schemas/SpecGenerationStatusEnum"
        analysisId:
          type: string
          format: uuid
          description: Analysis ID for tracking
        message:
          type: string
          description: Optional status message
    SpecGenerationStatusResponse:
      type: object
      required:
        - status
        - analysisId
      properties:
        status:
          $ref: "#/components/schemas/SpecGenerationStatusEnum"
        analysisId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
          description: When generation started
        completedAt:
          type: string
          format: date-time
          description: When generation completed (if completed)
        errorMessage:
          type: string
          description: Error message if generation failed
    SpecGenerationStatus:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/SpecGenerationStatusEnum"
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        errorMessage:
          type: string
    SpecGenerationStatusEnum:
      type: string
      enum:
        - pending
        - running
        - completed
        - failed
        - not_found
      description: |
        Generation status:
        - pending: Job queued but not started
        - running: AI generation in progress
        - completed: Document successfully generated
        - failed: Generation failed
        - not_found: No generation request exists

    # Usage Quota Schemas
...